+++ devolo_home_control_api/exceptions/device.py
@@ -1,0 +1,10 @@
+class MprmDeviceCommunicationError(Exception):
+    """ Communicating to a device via mPRM failed """
+
+
+class MprmDeviceNotFoundError(Exception):
+    """ A device like this was not found """
+
+
+class WrongElementError(Exception):
+    """ This element was not meant for this property. """