+++ devolo_home_control_api/exceptions/general.py
@@ -1,0 +1,6 @@
+class WrongCredentialsError(Exception):
+    """ Wrong credentials were used. """
+
+
+class WrongUrlError(Exception):
+    """ Wrong URL was used. """