+++ devolo_home_control_api/properties/consumption_property.py
@@ -4,7 +4,8 @@ from typing import Any
 from requests import Session
 
 from ..devices.gateway import Gateway
-from .property import Property, WrongElementError
+from ..exceptions.device import WrongElementError
+from .property import Property
 
 
 class ConsumptionProperty(Property):