+++ devolo_home_control_api/properties/settings_property.py
@@ -3,7 +3,8 @@ from typing import Any
 from requests import Session
 
 from ..devices.gateway import Gateway
-from .property import Property, WrongElementError
+from ..exceptions.device import WrongElementError
+from .property import Property
 
 
 class SettingsProperty(Property):