+++ tests/test_voltage_property.py
@@ -1,6 +1,6 @@
 import pytest
 
-from devolo_home_control_api.properties.property import WrongElementError
+from devolo_home_control_api.exceptions.device import WrongElementError
 from devolo_home_control_api.properties.voltage_property import VoltageProperty